src/requirements.txt
physrisk-lib==0.31.0